+++ data/tasks.txt
@@ -3,5 +3,5 @@ T | 0 | homework
 E | 0 | jog | 2019-10-15 1600 | 2019-10-15 1700
 E | 0 | add homework | 2000-10-22 1600 | 2000-10-22 1700
 D | 0 | return book | 2019-12-02 1800
-E | 0 | add homework | 2024-01-11 1200 | 2024-01-22 1200
+D | 0 | return book | 2019-12-02 1800
 D | 0 | return book | 2019-12-02 1800
