--- src/main/java/ExitCommand.java
@@ -1,11 +1,0 @@
-public class ExitCommand extends CommandExecutor {
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showGoodbye();
-    }
-
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
