+++ src/main/java/arts/Arts.java
@@ -1,3 +1,17 @@
+package arts;
+
+import arts.util.Parser;
+import arts.util.Storage;
+import arts.util.Ui;
+import arts.task.TaskList;
+import arts.enums.CommandType;
+import arts.command.AddDeadlineCommand;
+import arts.command.AddEventCommand;
+import arts.command.MarkCommand;
+import arts.command.UnmarkCommand;
+import arts.command.DeleteCommand;
+import arts.command.AddTodoCommand;
+
 import java.time.format.DateTimeFormatter;
 
 public class Arts {
