+++ src/main/java/arts/ArtsException.java
@@ -1,3 +1,5 @@
+package arts;
+
 public class ArtsException extends Exception {
     public ArtsException(String message) {
         super(message);
