+++ src/main/java/arts/command/Command.java
@@ -1,3 +1,7 @@
+package arts.command;
+
+import arts.ArtsException;
+
 public interface Command {
     void execute() throws ArtsException;
 }
