+++ src/main/java/arts/enums/CommandType.java
@@ -1,3 +1,5 @@
+package arts.enums;
+
 public enum CommandType {
     BYE, LIST, MARK, UNMARK, DELETE, TODO, DEADLINE, EVENT
 }
