+++ src/main/java/arts/task/Deadline.java
@@ -1,3 +1,7 @@
+package arts.task;
+
+import arts.task.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
