+++ src/main/java/arts/task/Event.java
@@ -1,3 +1,5 @@
+package arts.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
