+++ src/main/java/arts/task/Task.java
@@ -1,3 +1,7 @@
+package arts.task;
+
+import arts.ArtsException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
