+++ src/main/java/arts/task/TaskList.java
@@ -1,3 +1,5 @@
+package arts.task;
+
 import java.util.ArrayList;
 
 public class TaskList {
