+++ src/main/java/arts/task/Todo.java
@@ -1,3 +1,5 @@
+package arts.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
