+++ src/main/java/arts/util/Ui.java
@@ -1,3 +1,5 @@
+package arts.util;
+
 import java.util.Scanner;
 
 public class Ui {
